requirements.txt
fastapi
uvicorn
websockets
httpx
python-multipart
python-dotenv
pydantic
jinja2
pytest
pytest-asyncio
psutil